% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/enumerate.R
\name{enumapply}
\alias{enumapply}
\title{enumapply works like lapply, but also passes the index of the
current element to FUN as a second variable.}
\usage{
enumapply(X, FUN, ...)
}
\arguments{
\item{X}{a vector (atomic or list)}

\item{FUN}{function to apply. first input receives elements of X, the second the index, the next ones from ...}

\item{...}{optional arguments to FUN}
}
\value{
a list of same length as X
}
\description{
enumapply works like lapply, but also passes the index of the
current element to FUN as a second variable.
}
\examples{
enumapply(c("a", "b", "c"), function(x, i)c(x,i))
}

