% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/enumerate.R
\name{enumerate}
\alias{enumerate}
\title{To be used in a for loop, it adds the
iteration number to the object, that can be
accessed via i_()}
\usage{
enumerate(x)
}
\arguments{
\item{x}{object to iterate on}
}
\value{
list of objects with class enumerate
}
\description{
To be used in a for loop, it adds the
iteration number to the object, that can be
accessed via i_()
}
\examples{
random_numbers <- rnorm(10)
random_numbers
for (num in enumerate(random_numbers)){
  print(i_(num))
  print(num)
}
}

