% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{file.date}
\alias{file.date}
\title{returns the date of a file
multiple files can be given as an input}
\usage{
file.date(file)
}
\arguments{
\item{file}{a vector of file paths}
}
\value{
a vector of time stamps as POSIXct
}
\description{
returns the date of a file
multiple files can be given as an input
}

