% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/enumerate.R
\name{i_}
\alias{i_}
\title{Acces the iteration number of objects created with enumerate()}
\usage{
i_(x)
}
\arguments{
\item{x}{object}
}
\description{
Acces the iteration number of objects created with enumerate()
}

