% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/run_script.R
\name{run_script}
\alias{run_script}
\title{Run an existing script with a certain structure while
checking if computations are necessary.}
\usage{
run_script(file, outputFiles, inputFiles = NULL, parameters = NULL)
}
\arguments{
\item{file}{path of the .R script to run}

\item{outputFiles}{files to be created by the script}

\item{inputFiles}{files used as an input by the script}

\item{parameters}{additional parameters, as a list}
}
\description{
Run an existing script with a certain structure while
checking if computations are necessary.
}

