% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/run_script.R
\name{run_script}
\alias{run_script}
\title{Function to take over an existing scripts with standard parameters
to allow for standardized etl processing}
\usage{
run_script(file, outputFiles, inputFiles = NULL, parameters = NULL)
}
\arguments{
\item{file}{}

\item{outputFiles}{}

\item{inputFiles}{}

\item{parameters}{}
}
\description{
Function to take over an existing scripts with standard parameters
to allow for standardized etl processing
}

